+++ Sources/VCore/Services and Managers/Presentation Host/PresentationHostDataSourceCache.swift
@@ -1,0 +1,36 @@
+//
+//  PresentationHostDataSourceCache.swift
+//  VCore
+//
+//  Created by Vakhtang Kontridze on 26.05.22.
+//
+
+#if os(iOS)
+
+import SwiftUI
+
+// MARK: - Presentation Host Data Source Cache
+final class PresentationHostDataSourceCache {
+    // MARK: Properties
+    static let shared: PresentationHostDataSourceCache = .init()
+    
+    private var storage: [String: Any] = [:]
+    
+    // MARK: Initializers
+    private init() {}
+    
+    // MARK: Get and Set
+    func get(key: String) -> Any? {
+        storage[key]
+    }
+    
+    func set(key: String, value: Any) {
+        storage[key] = value
+    }
+    
+    func remove(key: String) {
+        storage[key] = nil
+    }
+}
+
+#endif